+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## xxxx-xx-xx - dev
 
 * [Spec] Clarify CRS usage expectation for geographic properties
-* [Deployment] Update docker base image
+* [Deployment] Update docker base image, move Dockerfile
 * [Deployment] Add `azure-pipelines.yml`
 
 ## 2025-10-30 - 0.2.4